advanced/capstone_project/solution/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-RESTX==1.1.0
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21

# Data processing and analysis
pandas==2.1.1
numpy==1.25.2
scipy==1.11.3

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Background tasks
celery==5.3.2
redis==5.0.0

# Security
bcrypt==4.0.1
cryptography==41.0.4

# File handling
openpyxl==3.1.2
xlrd==2.0.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-flask==1.2.0
pytest-mock==3.11.1
coverage==7.3.2

# Development tools
python-dotenv==1.0.0
marshmallow==3.20.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Production server (optional)
gunicorn==21.2.0